refactor(teacher): split StudentProgressTab + AssignmentTab into hooks + components (Related to #1828)
StudentProgressTab: 1178 → 303 lines
AssignmentTab: 1068 → 420 lines

New files:
- hooks/useStudentProgress.ts — data fetch, cache, tag/session/dialogue/curve state
- hooks/useAssignments.ts — data fetch, form state, edit/delete/expand state
- components/StudentProgressCard.tsx — mobile card view
- components/StudentExpandedPanel.tsx — expanded sessions + dialogue + chart
- components/StudentDialogueModal.tsx — dialogue history modal
- components/StudentTagManager.tsx — tag CRUD UI
- components/studentProgressUtils.tsx — shared display helpers
- components/AssignmentCard.tsx — mobile card view
- components/AssignmentCreateForm.tsx — create form panel
- components/AssignmentExpandedPanel.tsx — expanded detail
- components/AssignmentInlineEdit.tsx — inline edit form
- components/AssignmentStatusTabs.tsx — status filter tabs
- components/AssignmentEmptyState.tsx — empty state
- components/assignmentUtils.ts — formatDate, isOverdue, completionPercentage

Build: vite build ✓ zero TypeScript errors
