+++ frontend/src/pages/teacher/components/studentProgressUtils.tsx
@@ -1,0 +1,43 @@
+import React from 'react';
+
+export const TAG_COLOR_CLASSES: Record<string, string> = {
+  red: 'bg-red-100 text-red-700 border-red-200',
+  orange: 'bg-orange-100 text-orange-700 border-orange-200',
+  green: 'bg-green-100 text-green-700 border-green-200',
+  blue: 'bg-blue-100 text-blue-700 border-blue-200',
+  purple: 'bg-purple-100 text-purple-700 border-purple-200',
+  gray: 'bg-gray-100 text-gray-600 border-gray-200',
+};
+
+export function tagColorClass(color: string): string {
+  return TAG_COLOR_CLASSES[color] ?? TAG_COLOR_CLASSES.gray;
+}
+
+export function formatDate(dateStr: string | null): string {
+  if (!dateStr) return '-';
+  return new Date(dateStr).toLocaleDateString('zh-TW', {
+    year: 'numeric',
+    month: 'short',
+    day: 'numeric',
+  });
+}
+
+export function formatScore(score: number | null): string {
+  if (score === null || score === undefined) return '-';
+  return `${Math.round(score)}%`;
+}
+
+export function statusLabel(status: string) {
+  if (status === 'completed') {
+    return (
+      <span className="inline-block px-1.5 py-0.5 rounded text-xs font-medium bg-green-100 text-green-700">
+        已完成
+      </span>
+    );
+  }
+  return (
+    <span className="inline-block px-1.5 py-0.5 rounded text-xs font-medium bg-yellow-100 text-yellow-700">
+      進行中
+    </span>
+  );
+}
